requirements.txt
aioredis
discord.py @ git+https://github.com/Rapptz/discord.py@rewrite
jishaku
pillow
